+++ attributes/default.rb
@@ -1,10 +1,17 @@
 #
 # Cookbook Name:: taurus
-#
+
 default['java']['jdk_version'] = '7'
+default['erlang']['gui_tools'] = false
+default['erlang']['install_method'] = 'package'
+
 default['taurus']['jmeter_support'] = true
+default['taurus']['jmeter_service'] = true
 default['taurus']['locustio_support'] = true
 default['taurus']['ab_support'] = true
+default['taurus']['siege_support'] = true
+default['taurus']['gatling_support'] = true
+default['taurus']['tsung_support'] = true
 
 default['taurus']['user'] = 'taurus'
 default['taurus']['group'] = 'taurus'
@@ -27,10 +34,15 @@ default['taurus']['package']['list_ubuntu'] = ['python-dev', 'python-pip',
 default['taurus']['python']['list'] = %w(lxml psutil pyzmq gevent)
 
 # JMeter
-default['taurus']['jmeter']['version'] = '2.13' # 3.0
+default['taurus']['jmeter']['version'] = '3.0'
 default['taurus']['jmeter']['path'] = "#{node['taurus']['home']}/tools/jmeter"
 default['taurus']['jmeter']['mirror_source'] = 'https://archive.apache.org/dist/jmeter/binaries'
 default['taurus']['jmeter']['source_url'] = "#{node['taurus']['jmeter']['mirror_source']}/apache-jmeter-#{node['taurus']['jmeter']['version']}.zip"
+
+# JMeter Server (runit) for Distributed
+default['taurus']['jmeter']['server_rmi_port'] = '1098'
+default['taurus']['jmeter']['client_rmi_port'] = '1099'
+default['taurus']['jmeter']['log_dir'] = '/var/log/jmeter-service'
 
 # JMeter plugins
 default['taurus']['jmeter']['plugins']['version'] = '1.4.0'
@@ -41,10 +53,18 @@ default['taurus']['jmeter']['plugins']['mirror_source'] = 'http://jmeter-plugins
 default['taurus']['locustio']['version'] = '0.7.5'
 
 # Apache Benchmark
-default['taurus']['ab']['package']['list'] = ['httpd-tools']
-default['taurus']['ab']['package']['list_ubuntu'] = ['apache2-utils']
+default['taurus']['ab']['version'] = nil
 
-# TODO: Support additional performance tools
-# gatling
-# seige
-# tsung
+# Siege
+default['taurus']['siege']['version'] = nil
+
+# Gatling
+default['taurus']['gatling']['version'] = '2.1.7'
+default['taurus']['gatling']['path'] = "#{node['taurus']['home']}/tools/gatling"
+default['taurus']['gatling']['mirror_source'] = "https://repo1.maven.org/maven2/io/gatling/highcharts/gatling-charts-highcharts-bundle/#{node['taurus']['gatling']['version']}"
+default['taurus']['gatling']['source_url'] = "#{node['taurus']['gatling']['mirror_source']}/gatling-charts-highcharts-bundle-#{node['taurus']['gatling']['version']}-bundle.zip"
+
+# Tsung
+default['taurus']['tsung']['version'] = '1.6.0'
+default['taurus']['tsung']['mirror_source'] = 'http://tsung.erlang-projects.org/dist'
+default['taurus']['tsung']['source_url'] = "#{node['taurus']['tsung']['mirror_source']}/tsung-#{node['taurus']['tsung']['version']}.tar.gz"