+++ test/integration/default/serverspec/default_spec.rb
@@ -27,7 +27,9 @@ content = [
   '',
   'modules:',
   '  jmeter:',
-  '    path: /opt/taurus/tools/jmeter'
+  '    path: /opt/taurus/tools/jmeter',
+  '  gatling:',
+  '    path: /opt/taurus/tools/gatling'
 ].join("\n") << "\n"
 
 describe file('/opt/taurus/.bzt-rc') do
@@ -62,7 +64,7 @@ end
 
 describe command('/opt/taurus/tools/jmeter/bin/jmeter --version') do
   its(:exit_status) { should eq 0 }
-  its(:stdout) { should match %r{Version 2.13} }
+  its(:stdout) { should match %r{3.0} }
 end
 
 describe command('locust --version 2>&1') do
@@ -73,4 +75,29 @@ end
 describe command('ab -V 2>&1') do
   its(:exit_status) { should eq 0 }
   its(:stdout) { should match %r{ApacheBench, Version 2} }
+end
+
+describe command('siege --version 2>&1') do
+  its(:exit_status) { should eq 0 }
+  its(:stdout) { should match %r{SIEGE 3} }
+end
+
+describe file('/opt/taurus/tools/gatling') do
+  it { should be_directory }
+  it { should be_owned_by 'taurus' }
+  it { should be_mode 755 }
+end
+
+describe command('/opt/taurus/tools/gatling/bin/gatling.sh --help 2>&1') do
+  its(:exit_status) { should eq 0 }
+  its(:stdout) { should match %r{GATLING_HOME is set to /opt/taurus/tools/gatling} }
+end
+
+describe command('erl -eval \'erlang:display(erlang:system_info(otp_release)), halt().\'  -noshell 2>&1') do
+  its(:exit_status) { should eq 0 }
+end
+
+describe command('tsung -v 2>&1') do
+  its(:exit_status) { should eq 0 }
+  its(:stdout) { should match %r{Tsung version 1.6} }
 end