+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 taurus CHANGELOG
 ==============================
 
+# 2.5.0
+
+Add jmeter server support via runit. Updated documentation.
+
+# 2.4.2
+
+Hotfix for spec test, de-duplicate jmeter-plugin ark downloads (remove CHEF-3694 warnings)
+Cleanup package installation for tools and compact variables used.
+
+# 2.4.1
+
+Update documentation and include CI references
+
+# 2.4.0
+
+Add support for the Tsung Engine
+
+# 2.3.0
+
+Add support for the Gatling Engine
+
+# 2.2.0
+
+Add support for the Siege Engine
+
 # 2.1.2
 
 Fix platform selection for installation of yum-epel/apt